// app/pages/timelog/timelog.d.ts
import { Config, Nav, NavParams, ViewController } from 'ionic-angular';
import { TimeProvider } from '../../providers/time-provider';
export declare class TimelogPage {
    private nav;
    private navParams;
    private timeProvider;
    private config;
    private view;
    inc: number;
    isbillable: boolean;
    timecount: any;
    mintime: number;
    time: any;
    timenote: string;
    title: string;
    he: any;
    selects: any;
    displayFormat: string;
    minuteValues: Array<number>;
    start_time: string;
    stop_time: string;
    UserDateOffset: number;
    constructor(nav: Nav, navParams: NavParams, timeProvider: TimeProvider, config: Config, view: ViewController);
    decrement(): void;
    increment(): void;
    ngAfterViewInit(): void;
    AddHours(date: any, hours: any): any;
    ngOnInit(): void;
    saveSelect(event: any): void;
    onSubmit(form: any): void;
    setDate(date: any, showmonth?: any, istime?: any): any;
    setMinTime(date: any): any;
    setMaxTime(date: any): any;
    getStartDate(time: any): any;
    setStartDate(time: any): void;
    setStopDate(time: any): void;
    updateHours(): void;
    endsWith(str: any, search: any): boolean;
    getInterval(): number;
    getFixed(value: any): string;
    close(): void;
}